Added tracefile and bluetooth manger class. Breaked the RXData parser method.
